don't set the HOME environment variable for singularity/apptainer
When using the `--home` flag while running, singularity/apptainer
handle updating the in-container HOME environment variable and
intentionally prefer the path passed to `--home`. This means the path
set in these lines would be ignored. For our case, since they are the
same, it isn't a big deal and we just remove the HOME update to avoid
apptainer printing a warning.