manually install singularity from deb
this decides our singularity pin to be the last version of singularity
released by apptainer. We do not follow sylabs singularity (here
referred to as singularityCE). We simply download the deb of that
version and apt install it on ubuntu

this will hopefully help this test last longer since the eWaterCycle
action was written for Node 12 but forced to run on Node 16 which I
assume is error prone