patch invalid var name when using screen (#134)
`screen` creates an environment variable called `TERMCAP` which is split
across multiple lines and (I presume) stores some configuration details
of `screen`. While the first line of the `TERMCAP` value contains a
space and is therefore removed, subsequent lines are not removed by the
invalid-value regex and thus result in attempting to pass (for example)
```
        :DO=\E[%dB:LE=\E[%dD:RI=\E[%dC:UP=\E[%dA:bs:bt=\E[Z:\
```
as an environment variable. This gets injected all the way into
apptainer's environment injection script which breaks because it parses
`:DO` as the environment variable name and colon's are not allowed in
POSIX compliant shell environment variable names.

The patch is to also remove variables whose names contain `:` so that
they are not causing errors within the POSIX-sh inside of the container.