patch overly strict env var names (#137)
We just need to be more specific about how we handle colons. They cannot
end up within a variable name since that is not supported in shells and
thus breaks apptainer's environment injection script but we want to
allow them in variable values since they are often used to separate many
options. The compromise is to allow them in non-whitespace contexts and
have a special check when looking through all environment variables on
the host. (Looking at you screen's TERMCAP).