warn if denv init inside denv (#138)
When `denv init`ing, make sure to prompt the user if they are going to
overwrite (the workspace to-be-initialized already has a denv) or
override (a parent directory has a denv). This is helpful to at least
make sure that the user acknowledges that they are entering a
potentially confusing situation.