specialize denv init for script writers (#142)
This introduces four new command line flags to the `denv init` CLI
making it easier for script writers to avoid the user prompt if desired.
This follows the idea outlined in
https://github.com/tomeichlersmith/denv/pull/140#issuecomment-2392305916

`--[no-]mkdir` answers the prompt about if denv is allowed to create the workspace
directory if it doesn't exist.
`--[no-]over` answers the prompt about if denv is allowed to overwrite/override an
existing configuration

Besides the introduction of these flags, more tests were included to make sure
they are oprational. The manual and CLI help was updated accordingly,
and some cleanup was done to use true/false instead of 1/0 within variables
storing the results of these flags.