piggy-back on apptainer SIF image cache (#143)
This resolves #141

`apptainer` has an in-container environment variable
`APPTAINER_CONTAINER` that holds the full path to the SIF image that was
used to launch the container.[^1] This environment variable is not
documented, but is exactly what we want since it allows us to use
`apptainer` to cache a local copy of the SIF image built from an OCI
image and then we just launch it to retrieve the path to this cached
image.

[^1]:
https://github.com/apptainer/apptainer/blob/e6fc6f0d5862a372241a8bec50b980036e37975e/internal/pkg/runtime/launch/launcher_linux.go#L526

The other singularity flavors have a similarly named environment
variable `SINGULARITY_CONTAINER` that is also not documented
but produces the same behavior.

This avoids manually building the SIF image ourselves for each denv
and instead just uses cached SIF image that apptainer builds when
running from and OCI image name.

I think this is the "correct" way to do this as far as I can tell from looking through the
documentation and code. apptainer/singularity really don't want folks to
rely on the cache especially when running multiple containers from the
same image which makes sense and motivates me to update the
documentation accordingly.

add test that makes sure users can invalidate the cache and the
image is rebuilt seamlessly

update documentation: as pointed out in [apptainer's cache
documentation](https://apptainer.org/docs/user/1.3/build_env.html#sec-cache),
if the filesystem the cache is residing in does not support atomic
renaming, then running the image from that location may see performance
issues (reading between the lines here). I'd like to point this out in
the `denv` docs. `denv` supports running from a manually built SIF image
as well, so if users are preparing a cluster-wide processing campaign,
they should manually build and fix a SIF that they can then provide
`denv`.