use exec instead of run to avoid entrypoint (#147)
we dont want to use the image's entrypoint ever, we have our own and
this remains the case even when pulling the image. 

This only affects apptainer/singularity runners where we need to use the exec
subcommand to bypass the entrypoint rather than the run subcommand which
will respect the entrypoint.