make denv fail if unable to meet override necessity
denv now prints an error message and returns a non-zero status if the
user does not allow it to override (with --no-over) but denv needs to
override a denv present in a parent directory in order to form a
properly initialized denv within the workspace requested (PWD if not
provided on the command line).

denv will still quietly skip overWRITING because, in that case, the
requested workspace /is/ a properly initialized denv and the user has
other means they should use to alter its configuration (namely `denv
config`).