allow for denv name to be an invalid hostname (#155)
We allow for underscores in the name because they are often used within directory names which are the default names for the denv. Any other special character should trigger an error during `denv init` that informs the user of the issue in a more helpful manner than the (currently `apptainer`-only) error message from the underlying runner.